requirements.txt
# Core Dependencies
torch>=1.9.0
transformers>=4.12.0
pyspark>=3.2.0
kafka-python>=2.0.2
spacy>=3.1.0
nltk>=3.6.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Web Framework (for API endpoints)
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Monitoring
prometheus-client>=0.11.0
aiohttp>=3.8.0
psutil>=5.8.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-asyncio>=0.15.0

# Development Tools
jupyter>=1.0.0
ipykernel>=6.0.0
black>=21.7b0
flake8>=3.9.0
isort>=5.9.0

# Database
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=5.4.0
tqdm>=4.62.0
requests>=2.26.0
python-json-logger>=2.0.0

# Model Serialization
joblib>=1.0.0
dill>=0.3.4